requirements.txt
httpx~=0.23.3
icalendar~=5.0.4
pypdf~=3.4.0
Spire.Doc
